create header for intro_loader
